Revert "Merge branch 'CXOPS-2866' of github.com:crunchyroll/ef-open into CXOPS-2968"
This reverts commit 792697bb2d07a2a39647cd68acfbaa966ad87303, reversing
changes made to add5a83d01d2ebb86dc40db3f62a0aee9456f53a.
