+++ .gitignore
@@ -6,5 +6,3 @@
 .pants.workdir.file_lock
 .pids
 dist
-.coverage
-
